requirements.txt
torch
torchaudio
openai-whisper
transformers
sounddevice
numpy
scipy
pydub
gTTS
playsound
